% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_statistics.R
\name{boot_mean}
\alias{boot_mean}
\alias{boot_geometric}
\alias{boot_median}
\alias{boot_mode}
\alias{boot_prop}
\alias{boot_delta_mean}
\alias{boot_delta_geometric}
\alias{boot_delta_median}
\alias{boot_OR}
\title{`boot` statistic functions}
\usage{
boot_mean(d, i)

boot_geometric(d, i)

boot_median(d, i)

boot_mode(d, i)

boot_prop(d, i)

boot_delta_mean(d, i)

boot_delta_geometric(d, i)

boot_delta_median(d, i)

boot_OR(d, i)
}
\arguments{
\item{d}{A data frame.

   For `boot_mean`, `boot_geometric`, `boot_median`, and `boot_mode`, the data frame must consist of a single column of numeric data.

   For `boot_prop`, the data frame must consist of a single column of binomial categorical data.

   For `boot_delta_mean`, `boot_delta_geometric`, and `boot_delta_median`, the data frame must consist of a two columns. The first column must be numeric data, and the second column must be a two-level grouping variable.

   For `boot_OR` the data frame must consist of a two columns. The first column must be binomial categorical data, and the second column must be a two-level grouping variable.}

\item{i}{Internal indexing parameter used by \code{\link[boot]{boot}} for resampling.
\strong{Do not change}.}
}
\value{
Returns and object of class "boot".
}
\description{
A series of functions for internal use within \code{\link[boot]{boot}} only. The functions calculate measures of central tendency (arithmetic mean, geometric mean, median, and mode), differences in measures of central tendency (difference in means, difference in geometric means, difference in medians), proportions, and odds ratios.
}
\section{Functions}{
\itemize{
\item \code{boot_mean}: Calculates an arithmetic mean.

\item \code{boot_geometric}: Calculates a geometric mean

\item \code{boot_median}: Calculates a median

\item \code{boot_mode}: Calculates a mode

\item \code{boot_prop}: Calculates a proportion

\item \code{boot_delta_mean}: Calculates the difference between arithmetic means

\item \code{boot_delta_geometric}: Calculates the difference between geometric means

\item \code{boot_delta_median}: Calculates the difference between medians

\item \code{boot_OR}: Calculates an odds ratio
}}

