% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iris_reduced.R
\docType{data}
\name{iris_reduced}
\alias{iris_reduced}
\title{Simplified version of the 'iris' dataset.}
\format{A data frame data frame with 100 cases (rows) and 5 variables (columns):
\describe{
  \item{Sepal.Length}{Iris flower sepal length in centimeters.}
  \item{Sepal.Width}{Iris flower sepal width in centimeters.}
  \item{Petal.Length}{Iris flower petal length in centimeters.}
  \item{Petal.Width}{Iris flower petal width in centimeters.}
  \item{Species}{Species of iris in the dataset: \emph{iris setosa} and \emph{iris versicolor}.}
}}
\source{
\code{\link[datasets]{iris}}
}
\usage{
iris_reduced
}
\description{
This is a simplified version of the (Fisher or Anderson) \emph{iris} dataset.
}
\keyword{datasets}
